+++ pcs/test/test_cluster_pcmk_remote.py
@@ -68,8 +68,17 @@ class NodeAddRemote(ResourceTest):
                         />
                     </instance_attributes>
                     <operations>
+                        <op id="node-name-migrate_from-interval-0s"
+                            interval="0s" name="migrate_from" timeout="60"
+                        />
+                        <op id="node-name-migrate_to-interval-0s" interval="0s"
+                            name="migrate_to" timeout="60"
+                        />
                         <op id="node-name-monitor-interval-60s" interval="60s"
                             name="monitor" timeout="30"
+                        />
+                        <op id="node-name-reload-interval-0s" interval="0s"
+                            name="reload" timeout="60"
                         />
                         <op id="node-name-start-interval-0s" interval="0s"
                             name="start" timeout="60"