+++ CHANGELOG.md
@@ -16,9 +16,14 @@
 - CIB changes are now pushed to pacemaker as a diff in commands overhauled to
   the new architecture (previously the whole CIB was pushed). This resolves
   race conditions and ACLs related errors when pushing CIB. ([rhbz#1441673])
+- All actions / operations defined in resource agent's metadata (except
+  meta-data, status and validate-all) are now copied to the CIB when creating
+  a resource. ([rhbz#1418199], [ghissue#132])
 
 [ghissue#118]: https://github.com/ClusterLabs/pcs/issues/118
+[ghissue#132]: https://github.com/ClusterLabs/pcs/issues/132
 [rhbz#1388783]: https://bugzilla.redhat.com/show_bug.cgi?id=1388783
+[rhbz#1418199]: https://bugzilla.redhat.com/show_bug.cgi?id=1418199
 [rhbz#1420437]: https://bugzilla.redhat.com/show_bug.cgi?id=1420437
 [rhbz#1435697]: https://bugzilla.redhat.com/show_bug.cgi?id=1435697
 [rhbz#1441673]: https://bugzilla.redhat.com/show_bug.cgi?id=1441673